fix: 버그 및 디자인 수정
